That's correct! The commit message for this patch is:
This type of commit indicates that we have added a new feature to our project. In this case, we have updated the  file with new information about our coding activities. The message is clear and descriptive, focusing on what changed (the addition of new data) and why (to keep our data up-to-date).
Note that in a real project, you may want to use more specific commit types like  or .