Here is a possible concise commit message based on the provided git diff and modified files:
In this message, we highlight the main changes and affected functionality by mentioning the added lines of code in "yesalert.js", the updated files "RecipientView.tsx", "NewAlert.tsx", "index.ts", and "queries.ts". We also provide a clear description of the change using conventional commit types (chore, feat, fix, refactor, etc.). Finally, we list the modified files to ensure that everyone is on the same page regarding what has changed.