This commit updates the coding activity and adds new data to the README.md file. The  type indicates that this is a new feature or enhancement. The message provides more detail about what has changed, including the number of lines added and removed for each file.
This commit fixes an issue in the README.md file by updating the accuracy of the data displayed. The  type indicates that this is a bug fix or correction. The message provides more detail about what has changed and why it was necessary to make the update.

This commit simplifies the coding activity data by removing unnecessary lines and columns. The  type indicates that this is a code improvement or optimization. The message provides more detail about what has changed and why it was necessary to make the update.

This commit adds a new section to the README.md file, specifically the "Coding Activity by Hour" section. The  type indicates that this is an improvement or addition to the documentation. The message provides more detail about what has changed and why it was necessary to make the update.