This is a concise commit message based on the given Git diff and modified files:
In this message, we focus on the main changes and affected functionality. We use the conventional commit type , which is appropriate for making optimizations or improvements to existing code. We also provide a brief summary of the changes made and the reason for making them (in this case, improving Active Time by 10 minutes).
The message includes the following information:
* The type of change ()
* A brief summary of the changes made ()
* The affected files (2 modified files)
* The time taken for the commit (17 minutes)
Overall, this message provides a clear and concise description of the changes made and why they were made.