Here is an example of a well-written Git commit message:
This commit message clearly communicates what was added and why it was added. It also uses a conventional commit type, which helps other developers understand the purpose of the change.
Here is an example of a poorly written Git commit message:

This commit message does not clearly communicate what was added and why it was added. It also does not use a conventional commit type, which can make it difficult for other developers to understand the purpose of the change.