The commit message for this change log is:
This message clearly states what has changed and why. It also includes a brief summary of the changes made, which can help other developers understand the purpose of the commit. Additionally, it uses conventional commit types such as "update" to make the commit easier to read and understand for other developers.