This commit message is as follows:
In this message, the main change is that I have made a small correction to the code. The "fix" keyword indicates that this was an error in the code and it has been corrected.
The "readme" is the file I've made the change to. This commit message is short and informative, making it easy for other developers to understand what change was made and why.