The message should be:
This message clearly explains the changes made to the files and provides the appropriate information about the number of lines added and removed in each file. It also mentions the modified files and their respective changes. The use of conventional commit types (feat, fix, refactor, etc.) helps to make the message more meaningful and easier for other developers to understand.