Here's a concise commit message based on the provided git diff and modified files:
This commit message provides a brief summary of the changes made to the activity data in . The  type indicates that this is an administrative change, while the scope is , which refers to the specific file affected. The message also includes a list of modified functions and files, as well as the duration of the activity.