Here is a suggested message for the GitHub release:
This message provides a clear overview of the coding activity that took place on March 21st, 2025. It includes information about the lines of code added and deleted by file type, as well as the estimated activity count by hour. The use of conventional commit types (e.g., , , ) helps to ensure that the message is descriptive and easy to understand for those who are not familiar with the project's coding conventions.