The  command outputs a list of modified files and their associated changes. The output includes information about the number of lines added or removed, as well as the overall net change in lines and active time spent on each file. Additionally, the command can be used to generate visualizations such as pie charts and bar graphs that summarize the coding activity by file type and hour of day. Here is an example of how the  command could be used:
In this example, the  command is used to compare two versions of a README file and generate an output that summarizes the changes made between them. The output includes information about the number of lines added or removed, as well as the overall net change in lines and active time spent on each file. Additionally, the command can be used to generate visualizations such as pie charts and bar graphs that summarize the coding activity by file type and hour of day.