Sure, here's an example of a well-formatted commit message for a feature implementation:
In this example, we use the  type to indicate that a new feature has been added. We also provide a clear and concise description of the changes made in the commit message itself. The message includes the specific details of what was changed (i.e., the addition of a new "feature" functionality), which helps developers understand the purpose of the change.
Here's another example of a well-formatted commit message for a bug fix:

In this case, we use the  type to indicate that a bug has been fixed. We also provide a clear and concise description of the changes made (i.e., the resolution of an invalid data display issue). This helps developers understand the purpose of the change and how it affects the overall application functionality.
Remember, the commit message should be descriptive enough to help developers understand the purpose of the change, but not so long that it becomes overwhelming. Aim for a balance between being informative and concise.