The following is a concise commit message based on the given git diff and modified files:
This commit message follows the conventional commit type "chore" for small changes that don't require a major version bump. The scope of this change is "cda", as it updates the readme file for the cda project. The summary describes the main change made in the README file, which is an update to the statistics.
The message is brief and clear, focusing on what changed and why. It uses conventional commit types to indicate that this is a small change that doesn't require a major version bump.