Here is a possible concise commit message based on the given git diff and modified files:
This commit message follows the conventional commit format and clearly communicates that the changes made are related to updating activity statistics in the README file for the "cda" project on April 4, 2025. The use of the "refactor" type helps to signal that this is a non-trivial change that has been made to improve the codebase or make it more efficient, rather than introducing new features or fixing bugs.