Here's a concise commit message based on the provided git diff:
The commit message includes a clear description of the changes made, such as the addition and removal of lines in specific files, as well as the increase in active time due to additional changes. The type of the commit is "fix" to indicate that it fixes a bug or updates an existing feature. The scope is "cda" to indicate that the changes affect the cda project. Finally, the message includes a summary of the changes and any breaking changes introduced by the commit.