Here is a concise commit message based on the provided git diff and modified files:
In this message, we use the conventional commit type "feat" to indicate that we have made changes to an existing feature. We also provide a brief description of the changes made in the form of bullet points under the "Modified functions" and "Files changed" sections. The "Last Updated" section is updated to reflect the new time stamp.